locations/models.py 만 입력 users/model 복구
